pyproject.toml
[tool.poetry]
name = "django_default_project"
version = "0.8.0"
description = "Its not always good to start from scratch .. lets have some sane defaults .."
authors = ["Andy Grabow <andy@freilandkiwis.de>"]
license = "MIT"
readme = 'README.md'

repository = "https://github.com/kakulukia/django-default-project"
homepage = "https://github.com/kakulukia/django-default-project"

[tool.poetry.dependencies]
django = "^4.0"
django-secrets = "^1.0.2"
django-undeletable = "1.1.2"
django-extensions = "^3.1.1"
dj-static = "^0.0.6"
django-compressor = "^4.0"
django-axes = "^5.2.0"
raven = "^6.10.0"
pypugjs = "^5.8.1"
django-post-office = "^3.2.1"
uwsgidecorators = "^1.1.0"
huepy = "^1.2.1"
djangorestframework = "^3.11.0"
requests = "^2.22.0"
pendulum = "^2.0.5"
Pillow = "^9.0.1"
python = "^3.9.9"
sentry-sdk = "^1.5.7"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.261"
pre-commit = "^3.2.2"
icecream = "^2.1.3"
black = "^23.1.0"
django-debug-toolbar = "^3.2"
ipdb = "^0.13.6"
djdt-flamegraph = "^0.2.12"
fabric = "^3.0.0"
django-browser-reload = "^1.7.0"


[tool.black]
line-length = 99
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
/migrations/  # exclude django migrations
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 99
exclude = [
    "migrations",
    "node_modules",
    ".venv",
]
